pyproject.toml
[tool.poetry]
name = "final-test-2"
version = "0.1.0"
description = ""
authors = ["Landers"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
coverage = "^7.6.9"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
